requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
crewai==0.28.0
langchain==0.1.10
langchain-groq==0.1.3
chromadb==0.4.24
sentence-transformers==2.5.1
groq==0.4.2
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.3
tiktoken==0.5.2
numpy==1.26.4
